Split search page functionality into its own file
