+++ layouts/partials/footer-scripts.html
@@ -1,3 +1,7 @@
+{{ if eq .BaseFileName "search" }}
+    <script src="{{ (index $.Site.Data.manifest "search.js") | relURL}}"></script>
+{{ end }}
+
 {{ if .Params.footer_scripts }}
     {{ range $k, $v := .Params.footer_scripts }}
